Django_guide/requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
tzdata==2025.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
django-extensions==3.2.3
